renamed order.php to index.php and index.php to categories.php. and the links to get to the corresponding pages
