+++ adminpanel/newCategory.php
@@ -21,7 +21,7 @@ if (isset($_POST['save'])){
        
         $output = "<div class='newCategoryMessage'>
                         <p class='newCategoryMessage-text'>Denna kategori finns redan.</p>
-                        <a href='index.php'>OK</a>
+                        <a href='categories.php'>OK</a>
                         </div>";
 
         echo $output;
@@ -40,7 +40,7 @@ if (isset($_POST['save'])){
 
                         $output = "<div id='newCategoryMessage' class='newCategoryMessage'>
                         <p class='newCategoryMessage-text'>Kategorin är tillagt.</p>
-                        <a href='index.php#category'>OK</a>
+                        <a href='categories.php#category'>OK</a>
                         </div>";
                    
         echo $output;
