+++ adminpanel/newProduct.php
@@ -3,7 +3,7 @@
 require_once "db.php";
 require_once 'header.php';
 
-$ids = isset($_GET['category_id']) ? $_GET['category_id'] : header('Location:index.php');
+$ids = isset($_GET['category_id']) ? $_GET['category_id'] : header('Location:categories.php');
 
 if ($_SERVER['REQUEST_METHOD'] === 'POST') :
